requirements.txt
cryptography
keyring
pillow
pyotp
pyzbar
